# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cultivar Collection Management API",
    "description": "Advanced cannabis cultivar tracking and analytics platform",
    "version": "2.0.0",
    "x-generated": {
      "timestamp": "2025-11-01T19:55:58.334401",
      "source": "app.fastapi_app",
      "generator": "scripts/generate_openapi.py"
    }
  },
  "paths": {
...
    "/ws/broadcast": {
      "post": {
        "tags": [
          "WebSocket"
        ],
        "summary": "Broadcast Message",
        "description": "Admin endpoint to broadcast messages to all connected clients.",
        "operationId": "broadcast_message_ws_broadcast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Message"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}